Functiontest remake & Data rename
Functiontest now changeable.
Renamed main.php to MainInterface.php.
main.php for function call.
